chapter 3 exercises, labs, your turn, and chapter 4 your turn progress
I apologize this commit is a combination of a couple of commits because their was a large file accidentally downloaded onto my html_css folder and it caused it to stop pushing commits to orgin and would stay only locally. I am deeply sorry as i misled you on discord.